refactored : network_interface
