+++ libp2p/host/routed_host.py
@@ -1,11 +1,11 @@
+from libp2p.abc import (
+    INetworkService,
+)
 from libp2p.host.basic_host import (
     BasicHost,
 )
 from libp2p.host.exceptions import (
     ConnectionFailure,
-)
-from libp2p.network.network_interface import (
-    INetworkService,
 )
 from libp2p.peer.peerinfo import (
     PeerInfo,
